- Refactors navbar - Creates NavbarDesktop and NavbarMobile components.
